Removed hook useIsDesktop. Created helper getBreakpoint.
