+++ src/helpers/getBreakpoint.js
@@ -1,24 +1,13 @@
-function getBreakpoint(value = '') {
-  if (typeof value === 'number') {
-    return value;
+const getBreakpoint = function(width) {
+  if (width >= 992) {
+    return 'desktop';
   }
 
-  switch (value) {
-    case 'xs':
-      return 0;
-    case 'sm':
-      return 576;
-    case 'md':
-      return 768;
-    case 'lg':
-      return 992;
-    case 'xl':
-      return 1200;
-    case 'xxl':
-      return 1900;
-    default:
-      return 992;
+  if (width >= 768) {
+    return 'tablet';
   }
-}
+
+  return 'mobile';
+};
 
 export default getBreakpoint;
