+++ src/hooks/index.js
@@ -1,6 +1,5 @@
 import useGetBreakpoint from './useGetBreakpoint';
-import useIsDesktop from './useIsDesktop';
 import useIsOpened from './useIsOpened';
 import useQueryStaticImages from './useQueryStaticImages';
 
-export { useGetBreakpoint, useIsDesktop, useIsOpened, useQueryStaticImages };
+export { useGetBreakpoint, useIsOpened, useQueryStaticImages };
