+++ src/hooks/useGetBreakpoint.js
@@ -1,19 +1,6 @@
 import { useState, useEffect } from 'react';
 
-import { debounce } from '@helpers';
-
-// TODO: move to helpers -->
-const getBreakpoint = function(width) {
-  if (width >= 992) {
-    return 'desktop';
-  }
-
-  if (width >= 768) {
-    return 'tablet';
-  }
-
-  return 'mobile';
-};
+import { debounce, getBreakpoint } from '@helpers';
 
 const useGetBreakpoint = () => {
   const [breakpoint, setBreakpoint] = useState(null);
